move routes and queries out of server
